+++ importaciones/WEB-INF/classes/modules/import/classes/map/ClientQuotationItemMapBuilder.php
@@ -68,7 +68,7 @@ class ClientQuotationItemMapBuilder implements MapBuilder {
 
 		$tMap->addForeignKey('PRODUCTID', 'Productid', 'INTEGER', 'import_product', 'ID', true, null);
 
-		$tMap->addColumn('PRICE', 'Price', 'INTEGER', false, null);
+		$tMap->addColumn('PRICE', 'Price', 'FLOAT', false, null);
 
 		$tMap->addColumn('QUANTITY', 'Quantity', 'INTEGER', false, null);
